fix: extract AiThinkingThread previews to stay under 500-line cap
